Fix call to buienalarm and add sensor type
The call to Buienalarm would always use a timeframe of 60 (default) in stead of the provided timeframe. So I changed the calling to use keyword arguments in stead of positional arguments.

I have added a sensor type that will tell me if, and when, within the current timeframe rain is expected. -1 means no rain is expected, 0 means it is raining now and >0 means rain is expected within that amount of minutes